access validation - users

+++ app/Http/Controllers/Frontend/AgentListingScheduleController.php
@@ -13,24 +13,37 @@ use App\Models\Listing;
 use App\Models\ListingSchedule;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 
 
 class AgentListingScheduleController extends Controller
 {
-    public function index(AgentListingScheduleDataTable $listingScheduleDataTable, string $listing_id): View | JsonResponse
+    public function index(AgentListingScheduleDataTable $listingScheduleDataTable, string $listing_id): View | JsonResponse | RedirectResponse
     {
+        $listing = Listing::select('id', 'title', 'user_id')->where('id', $listing_id)->first();
+        if ($listing->user_id !== Auth::user()->id) {
+            return redirect()->back()->with('error', 'You are not authorized to edit this listing');
+        }
+        $listingTitle = $listing;
         $listingScheduleDataTable->with('listing_id', $listing_id);
-        $listingTitle = Listing::select('title')->where('id', $listing_id)->first();
         return $listingScheduleDataTable->render('frontend.dashboard.listing.schedule.index', compact('listing_id', 'listingTitle'));
     }
 
-    function create(Request $request, string $listing_id): View
+    function create(Request $request, string $listing_id): View | RedirectResponse
     {
+        $listing = Listing::select('id', 'user_id')->where('id', $listing_id)->first();
+        if ($listing->user_id !== Auth::user()->id) {
+            return redirect()->back()->with('error', 'You are not authorized to edit this listing');
+        }
         return view('frontend.dashboard.listing.schedule.create', compact('listing_id'));
     }
 
     function store(ListingScheduleStoreRequest $request, string $listing_id): RedirectResponse
     {
+        $listing = Listing::select('id', 'user_id')->where('id', $listing_id)->first();
+        if ($listing->user_id !== Auth::user()->id) {
+            return redirect()->back()->with('error', 'You are not authorized to edit this listing');
+        }
         $schedule = new ListingSchedule();
         $schedule->listing_id = $listing_id;
         $schedule->day = $request->day;
@@ -42,15 +55,23 @@ class AgentListingScheduleController extends Controller
         return to_route('user.listing-schedule.index', $listing_id)->with('success', 'Schedule has been added successfully');
     }
 
-    function edit(string $id): View
+    function edit(string $id): View | RedirectResponse
     {
         $schedule = ListingSchedule::findOrFail($id);
+        $listing = Listing::select('id', 'title', 'user_id')->where('id', $schedule->listing_id)->first();
+        if ($listing->user_id !== Auth::user()->id) {
+            return redirect()->back()->with('error', 'You are not authorized to edit this listing');
+        }
         return view('frontend.dashboard.listing.schedule.edit', compact('schedule'));
     }
 
     function update(ListingScheduleUpdateRequest $request, string $id): RedirectResponse
     {
         $schedule = ListingSchedule::find($id);
+        $listing = Listing::select('id', 'title', 'user_id')->where('id', $schedule->listing_id)->first();
+        if ($listing->user_id !== Auth::user()->id) {
+            return redirect()->back()->with('error', 'You are not authorized to edit this listing');
+        }
         $schedule->day = $request->day;
         $schedule->start_time = $request->start_time;
         $schedule->end_time = $request->end_time;
@@ -60,10 +81,14 @@ class AgentListingScheduleController extends Controller
         return to_route('user.listing-schedule.index', $schedule->listing_id)->with('success', 'Schedule has been updated successfully');
     }
 
-    function destroy(string $id): Response
+    function destroy(string $id): Response | RedirectResponse
     {
+        $schedule = ListingSchedule::findOrFail($id);
+        $listing = Listing::select('id', 'title', 'user_id')->where('id', $schedule->listing_id)->first();
+        if ($listing->user_id !== Auth::user()->id) {
+            return redirect()->back()->with('error', 'You are not authorized to edit this listing');
+        }
         try {
-            $schedule = ListingSchedule::findOrFail($id);
             $schedule->delete();
             return response(['status' => 'success', 'message' => 'Schedule has been deleted successfully']);
         } catch (\Exception $e) {
